move document reducer into core
